requirements.txt
fastapi
uvicorn[standard]
atlassian-python-api
beautifulsoup4
sumy
pydantic
